Fixed bug with coordinate system.
The X and Y co-ordinates for the tiles were swapped which was causing
tiles to have invalid references since the calculations for their
references to other tiles was based on a non swapped co-ordinate
system.

The boundary checking also had a bug where I was not checking for
values equal to the height and width of the grid.

These are fixed and everything seems to be in working order now.