Fixed memory leaks
The trie requires the length of the string when searching for a word.  For this
strlen was used, but because incomplete strings are searched for they are
non-null terminated. This means that strlen would search beyond the valid
length of the string and perform invalid reads of memory locations.  To fix
this, the current depth of traversal is being passed as it accurately
represents the number of characters of the current string.

Finally, when unloading the lexis and thus freeing the trie, TrieMakeEmpty was
the function being called. While this function does indeed free the tree, it
also initializes an empty tree. The correct function, TrieFree, is now being
called with correctly frees the tree completely.

With these two changes, blitzkrieg is currently free of memory leaks.